a2a/common/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a_common"
version = "0.1.0"
description = "Common utilities for A2A client interactions"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "httpx >= 0.20.0",
    "httpx-sse >= 0.4.0",
    "pyjwt >= 2.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/yourproject" # TODO: Update URL

[tool.setuptools.packages.find]
where = ["src"] 